@@ -4,79 +4,76 @@ namespace GoetasWebservices\Xsd\XsdToPhp\Php\Structure;
 
 class PHPArg
 {
-    protected $doc;
+    protected ?string $doc = null;
 
-    protected $type;
+    protected ?PHPClass $type = null;
 
-    protected $name;
+    protected ?string $name = null;
 
-    protected $nullable = false;
-    
-    protected $default;
+    protected bool $nullable = false;
 
-    public function __construct($name = null, $type = null)
+    protected string|array|null $default = null;
+
+    public function __construct(?string $name = null, ?PHPClass $type = null)
     {
         $this->name = $name;
         $this->type = $type;
     }
 
-    public function getDoc()
+    public function getDoc(): ?string
     {
         return $this->doc;
     }
 
-    public function setDoc($doc)
+    public function setDoc(?string $doc): static
     {
         $this->doc = $doc;
 
         return $this;
     }
 
-    /**
-     * @return PHPClass
-     */
-    public function getType()
+    public function getType(): ?PHPClass
     {
         return $this->type;
     }
 
-    public function setType(PHPClass $type)
+    public function setType(?PHPClass $type): PHPArg
     {
         $this->type = $type;
 
         return $this;
     }
 
-    public function getName()
+    public function getName(): ?string
     {
         return $this->name;
     }
 
-    public function setName($name)
+    public function setName(?string $name): PHPArg
     {
         $this->name = $name;
 
         return $this;
     }
 
-    public function getNullable()
+    public function getNullable(): bool
     {
         return $this->nullable;
     }
 
-    public function setNullable($nullable)
+    public function setNullable(bool $nullable): PHPArg
     {
         $this->nullable = $nullable;
-        
+
         return $this;
     }
-    
-    public function getDefault()
+
+    public function getDefault(): string|array|null
     {
         return $this->default;
     }
 
-    public function setDefault($default)
+    public function setDefault(string|array|null $default): PHPArg
     {
         $this->default = $default;
 
