@@ -4,15 +4,37 @@ namespace GoetasWebservices\Xsd\XsdToPhp\Php\Structure;
 
 class PHPClass
 {
-    protected $name;
+    protected ?string $name;
 
-    protected $namespace;
+    protected ?string $namespace;
 
-    protected $doc;
+    protected ?string $doc = null;
 
-    protected $implements = [];
+    protected array $implements = [];
 
-    public static function createFromFQCN($className)
+    protected array $checks = [];
+
+    /**
+     * @var PHPConstant[]
+     */
+    protected array $constants = [];
+
+    /**
+     * @var PHPProperty[]
+     */
+    protected array $properties = [];
+
+    protected bool $abstract = false;
+
+    protected ?PHPClass $extends = null;
+
+    public function __construct(?string $name = null, ?string $namespace = null)
+    {
+        $this->name = $name;
+        $this->namespace = $namespace;
+    }
+
+    public static function createFromFQCN(string $className): static
     {
         if (($pos = strrpos($className, '\\')) !== false) {
             return new self(substr($className, $pos + 1), substr($className, 0, $pos));
@@ -21,20 +43,12 @@ class PHPClass
         }
     }
 
-    /**
-     * @return array
-     */
     public function getImplements(): array
     {
         return $this->implements;
     }
 
-    /**
-     * @param bool $onlyParent
-     *
-     * @return PHPProperty
-     */
-    public function isSimpleType($onlyParent = false)
+    public function isSimpleType(bool $onlyParent = false): ?PHPProperty
     {
         if ($onlyParent) {
             $e = $this->getExtends();
@@ -48,14 +62,16 @@ class PHPClass
                 return $this->getPropertyInHierarchy('__value');
             }
         }
+
+        return null;
     }
 
-    public function setImplements(array $fqcn)
+    public function setImplements(array $fqcn): void
     {
         $this->implements = $fqcn;
     }
 
-    public function getPhpType()
+    public function getPhpType(): string
     {
         if (!$this->getNamespace()) {
             if ($this->isNativeType()) {
@@ -68,7 +84,7 @@ class PHPClass
         return '\\' . $this->getFullName();
     }
 
-    public function isNativeType()
+    public function isNativeType(): bool
     {
         return !$this->getNamespace() && in_array($this->getName(), [
             'string',
@@ -82,92 +98,58 @@ class PHPClass
         ]);
     }
 
-    public function __construct($name = null, $namespace = null)
-    {
-        $this->name = $name;
-        $this->namespace = $namespace;
-    }
-
-    public function getName()
+    public function getName(): ?string
     {
         return $this->name;
     }
 
-    public function setName($name)
+    public function setName(?string $name): static
     {
         $this->name = $name;
 
         return $this;
     }
 
-    public function getNamespace()
+    public function getNamespace(): ?string
     {
         return $this->namespace;
     }
 
-    public function setNamespace($namespace)
+    public function setNamespace(?string $namespace): static
     {
         $this->namespace = $namespace;
 
         return $this;
     }
 
-    public function getDoc()
+    public function getDoc(): ?string
     {
         return $this->doc;
     }
 
-    public function setDoc($doc)
+    public function setDoc(string $doc): static
     {
         $this->doc = $doc;
 
         return $this;
     }
 
-    public function __toString()
+    public function __toString(): string
     {
         return $this->getFullName();
     }
 
-    public function getFullName()
+    public function getFullName(): string
     {
         return "{$this->namespace}\\{$this->name}";
     }
 
-    protected $checks = [];
-
-    /**
-     * @var PHPConstant[]
-     */
-    protected $constants = [];
-
-    /**
-     * @var PHPProperty[]
-     */
-    protected $properties = [];
-
-    /**
-     * @param
-     *            $property
-     *
-     * @return array
-     */
-    public function getChecks($property)
+    public function getChecks(string $property): array
     {
-        return isset($this->checks[$property]) ? $this->checks[$property] : [];
+        return $this->checks[$property] ?? [];
     }
 
-    /**
-     * @param
-     *            $property
-     * @param
-     *            $check
-     * @param
-     *            $value
-     *
-     * @return $this
-     */
-    public function addCheck($property, $check, $value)
+    public function addCheck(string $property, string $check, mixed $value): static
     {
         $this->checks[$property][$check][] = $value;
 
@@ -177,56 +159,36 @@ class PHPClass
     /**
      * @return PHPProperty[]
      */
-    public function getProperties()
+    public function getProperties(): array
     {
         return $this->properties;
     }
 
-    /**
-     * @param string $name
-     *
-     * @return bool
-     */
-    public function hasProperty($name)
+    public function hasProperty(string $name): bool
     {
         return isset($this->properties[$name]);
     }
 
-    /**
-     * @param string $name
-     *
-     * @return bool
-     */
-    public function hasPropertyInHierarchy($name)
+    public function hasPropertyInHierarchy(string $name, bool $onlyParents = false): bool
     {
-        if (count($this->getProperties()) > 1 || (count($this->getProperties()) > 0 && !$this->hasProperty($name))) {
-            return false;
-        }
-        if ($this->hasProperty($name)) {
+        if (!$onlyParents && $this->hasProperty($name)) {
             return true;
         }
-        if (($this instanceof PHPClass) && $this->getExtends() && $this->getExtends()->hasPropertyInHierarchy($name)) {
+
+        if ($this->getExtends() && $this->getExtends()->hasPropertyInHierarchy($name)) {
             return true;
         }
 
         return false;
     }
 
-    /**
-     * @param string $name
-     *
-     * @return PHPProperty
-     */
-    public function getPropertyInHierarchy($name)
+    public function getPropertyInHierarchy(string $name, bool $onlyParents = false): ?PHPProperty
     {
-        if (count($this->getProperties()) > 1 || (count($this->getProperties()) > 0 && !$this->hasProperty($name))) {
-            return null;
-        }
-
-        if ($this->hasProperty($name)) {
+        if (!$onlyParents && $this->hasProperty($name)) {
             return $this->getProperty($name);
         }
-        if (($this instanceof PHPClass) && $this->getExtends() && $this->getExtends()->hasPropertyInHierarchy($name)) {
+
+        if ($this->getExtends() && $this->getExtends()->hasPropertyInHierarchy($name)) {
             return $this->getExtends()->getPropertyInHierarchy($name);
         }
 
@@ -234,11 +196,9 @@ class PHPClass
     }
 
     /**
-     * @param string $name
-     *
-     * @return PHPProperty
+     * @return PHPProperty[]
      */
-    public function getPropertiesInHierarchy()
+    public function getPropertiesInHierarchy(): array
     {
         $ps = $this->getProperties();
 
@@ -249,62 +209,38 @@ class PHPClass
         return $ps;
     }
 
-    /**
-     * @param string $name
-     *
-     * @return PHPProperty
-     */
-    public function getProperty($name)
+    public function getProperty(string $name): ?PHPProperty
     {
-        return $this->properties[$name];
+        return $this->properties[$name] ?? null;
     }
 
-    /**
-     * @return $this
-     */
-    public function addProperty(PHPProperty $property)
+    public function addProperty(PHPProperty $property): static
     {
         $this->properties[$property->getName()] = $property;
 
         return $this;
     }
 
-    /**
-     * @var bool
-     */
-    protected $abstract;
-
-    /**
-     * @var PHPClass
-     */
-    protected $extends;
-
-    /**
-     * @return PHPClass
-     */
-    public function getExtends()
+    public function getExtends(): ?PHPClass
     {
         return $this->extends;
     }
 
-    /**
-     * @return PHPClass
-     */
-    public function setExtends(PHPClass $extends)
+    public function setExtends(PHPClass $extends): static
     {
         $this->extends = $extends;
 
         return $this;
     }
 
-    public function getAbstract()
+    public function getAbstract(): bool
     {
         return $this->abstract;
     }
 
-    public function setAbstract($abstract)
+    public function setAbstract(bool $abstract): static
     {
-        $this->abstract = (bool) $abstract;
+        $this->abstract = $abstract;
 
         return $this;
     }
