@@ -2,29 +2,33 @@
 
 namespace GoetasWebservices\Xsd\XsdToPhp\Php\Structure;
 
+use Laminas\Code\Generator\AbstractMemberGenerator;
+
 class PHPProperty extends PHPArg
 {
-    /**
-     * @var string
-     */
-    protected $visibility = 'protected';
+    protected string $visibility = AbstractMemberGenerator::VISIBILITY_PROTECTED;
+    protected ?string $fixed = null;
 
-    /**
-     * @return string
-     */
-    public function getVisibility()
+    public function getVisibility(): string
     {
         return $this->visibility;
     }
 
-    /**
-     * @param string $visibility
-     *
-     * @return $this
-     */
-    public function setVisibility($visibility)
+    public function setVisibility(string $visibility): static
     {
         $this->visibility = $visibility;
+
+        return $this;
+    }
+
+    public function getFixed(): ?string
+    {
+        return $this->fixed;
+    }
+
+    public function setFixed(?string $fixed): static
+    {
+        $this->fixed = $fixed;
 
         return $this;
     }
