@@ -3,7 +3,7 @@
 namespace GoetasWebservices\Xsd\XsdToPhp\Php;
 
 use Exception;
-use GoetasWebservices\XML\XSDReader\Schema\Attribute\AttributeItem;
+use GoetasWebservices\XML\XSDReader\Schema\Attribute\AttributeSingle;
 use GoetasWebservices\XML\XSDReader\Schema\Attribute\Group as AttributeGroup;
 use GoetasWebservices\XML\XSDReader\Schema\Element\Element;
 use GoetasWebservices\XML\XSDReader\Schema\Element\ElementDef;
@@ -12,6 +12,9 @@ use GoetasWebservices\XML\XSDReader\Schema\Element\ElementRef;
 use GoetasWebservices\XML\XSDReader\Schema\Element\ElementSingle;
 use GoetasWebservices\XML\XSDReader\Schema\Element\Group;
 use GoetasWebservices\XML\XSDReader\Schema\Element\Choice;
+use GoetasWebservices\XML\XSDReader\Schema\Element\GroupRef;
+use GoetasWebservices\XML\XSDReader\Schema\Element\InterfaceSetFixed;
+use GoetasWebservices\XML\XSDReader\Schema\Element\InterfaceSetMinMax;
 use GoetasWebservices\XML\XSDReader\Schema\Element\Sequence;
 use GoetasWebservices\XML\XSDReader\Schema\Item;
 use GoetasWebservices\XML\XSDReader\Schema\Schema;
@@ -29,6 +32,10 @@ use Psr\Log\LoggerInterface;
 
 class PhpConverter extends AbstractConverter
 {
+    private array $classes = [];
+
+    private array $skipByType = [];
+
     public function __construct(NamingStrategy $namingStrategy, LoggerInterface $loggerInterface = null)
     {
         parent::__construct($namingStrategy, $loggerInterface);
@@ -53,9 +60,7 @@ class PhpConverter extends AbstractConverter
         });
     }
 
-    private $classes = [];
-
-    public function convert(array $schemas)
+    public function convert(array $schemas): array
     {
         $visited = [];
         $this->classes = [];
@@ -69,7 +74,7 @@ class PhpConverter extends AbstractConverter
     /**
      * @return PHPClass[]
      */
-    private function getTypes()
+    private function getTypes(): array
     {
         uasort($this->classes, function ($a, $b) {
             return strcmp($a['class']->getFullName(), $b['class']->getFullName());
@@ -84,11 +89,12 @@ class PhpConverter extends AbstractConverter
         return $ret;
     }
 
-    private function navigate(Schema $schema, array &$visited)
+    private function navigate(Schema $schema, array &$visited): void
     {
         if (isset($visited[spl_object_hash($schema)])) {
             return;
         }
+
         $visited[spl_object_hash($schema)] = true;
 
         foreach ($schema->getTypes() as $type) {
@@ -105,7 +111,7 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private function visitTypeBase(PHPClass $class, Type $type)
+    private function visitTypeBase(PHPClass $class, Type $type): void
     {
         $class->setAbstract($type->isAbstract());
 
@@ -119,50 +125,57 @@ class PhpConverter extends AbstractConverter
             $this->visitComplexType($class, $type);
         }
     }
-	
-	/**
-	 * Process xsd:complexType xsd:sequence xsd:element
-	 *
-	 * @param PHPClass $class
-	 * @param Schema $schema
-	 * @param Sequence $sequence
-	 */
-	private function visitSequence(PHPClass $class, Schema $schema, Sequence $sequence)
-	{
-		foreach ($sequence->getElements() as $childSequence) {
-			if ($childSequence instanceof Group) {
-				$this->visitGroup($class, $schema, $childSequence);
-			} else {
-				$property = $this->visitElement($class, $schema, $childSequence);
-				$class->addProperty($property);
-			}
-		}
-	}
-	
+
     /**
-     * Process xsd:complexType xsd:choice xsd:element
-     *
-     * @param PHPClass $class
-     * @param Schema $schema
-     * @param Choice $choice
+     * Process xsd:complexType xsd:sequence xsd:element
      */
-    private function visitChoice(PHPClass $class, Schema $schema, Choice $choice)
+    private function visitSequence(PHPClass $class, Schema $schema, Sequence $sequence): void
     {
-        foreach ($choice->getElements() as $choiceOption) {
-			if ($choiceOption instanceof Sequence) {
-				$this->visitSequence($class, $schema, $choiceOption);
-			} else {
-				$property = $this->visitElement($class, $schema, $choiceOption);
-				$class->addProperty($property);
-			}
+        foreach ($sequence->getElements() as $childSequence) {
+            if ($childSequence instanceof Group) {
+                $this->visitGroup($class, $schema, $childSequence);
+            } else {
+                $property = $this->visitElement($class, $schema, $childSequence);
+                $class->addProperty($property);
+            }
         }
     }
-    
-    private function visitGroup(PHPClass $class, Schema $schema, Group $group)
+
+    /**
+     * Process xsd:complexType xsd:choice xsd:element
+     */
+    private function visitChoice(PHPClass $class, Schema $schema, Choice $choice, ?GroupRef $groupRef = null): void
+    {
+        foreach ($choice->getElements() as $choiceOption) {
+            if ($choiceOption instanceof Sequence) {
+                $this->visitSequence($class, $schema, $choiceOption);
+            } elseif ($choiceOption instanceof Choice) {
+                $this->visitChoice($class, $schema, $choiceOption);
+            } elseif ($choiceOption instanceof Group) {
+                $this->visitGroup($class, $schema, $choiceOption);
+            } else {
+                /** @var Element $choiceOption */
+                if ($groupRef !== null) {
+                    $choiceOption->setMax($groupRef->getMax());
+                    $choiceOption->setMin($groupRef->getMin());
+                }
+                $property = $this->visitElement($class, $schema, $choiceOption);
+                $class->addProperty($property);
+            }
+        }
+    }
+
+    private function visitGroup(PHPClass $class, Schema $schema, Group $group): void
     {
         foreach ($group->getElements() as $childGroup) {
             if ($childGroup instanceof Group) {
                 $this->visitGroup($class, $schema, $childGroup);
+            } elseif ($childGroup instanceof Choice) {
+                $childGroupRef = null;
+                if ($group instanceof GroupRef) {
+                    $childGroupRef = $group;
+                }
+                $this->visitChoice($class, $schema, $childGroup, $childGroupRef);
             } else {
                 $property = $this->visitElement($class, $schema, $childGroup);
                 $class->addProperty($property);
@@ -170,7 +183,7 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private function visitAttributeGroup(PHPClass $class, Schema $schema, AttributeGroup $att)
+    private function visitAttributeGroup(PHPClass $class, Schema $schema, AttributeGroup $att): void
     {
         foreach ($att->getAttributes() as $childAttr) {
             if ($childAttr instanceof AttributeGroup) {
@@ -182,14 +195,7 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private $skipByType = [];
-
-    /**
-     * @param bool $skip
-     *
-     * @return PHPClass
-     */
-    public function visitElementDef(ElementDef $element)
+    public function visitElementDef(ElementDef $element): PHPClass
     {
         if (!isset($this->classes[spl_object_hash($element)])) {
             $schema = $element->getSchema();
@@ -200,7 +206,10 @@ class PhpConverter extends AbstractConverter
             $class->setDoc($element->getDoc());
 
             if (!isset($this->namespaces[$schema->getTargetNamespace()])) {
-                throw new Exception(sprintf("Can't find a PHP namespace to '%s' namespace", $schema->getTargetNamespace()));
+                throw new Exception(sprintf(
+                    "Can't find a PHP namespace to '%s' namespace",
+                    $schema->getTargetNamespace()
+                ));
             }
             $class->setNamespace($this->namespaces[$schema->getTargetNamespace()]);
 
@@ -223,12 +232,12 @@ class PhpConverter extends AbstractConverter
         return $this->classes[spl_object_hash($element)]['class'];
     }
 
-    public function isSkip($class)
+    public function isSkip($class): bool
     {
         return !empty($this->skipByType[spl_object_hash($class)]);
     }
 
-    private function findPHPName(Type $type)
+    private function findPHPName(Type $type): array
     {
         $schema = $type->getSchema();
 
@@ -260,14 +269,9 @@ class PhpConverter extends AbstractConverter
     }
 
     /**
-     * @param bool $force
-     * @param bool $skip
-     *
-     * @return PHPClass
-     *
      * @throws Exception
      */
-    public function visitType(Type $type, $force = false)
+    public function visitType(Type $type, bool $force = false): PHPClass
     {
         if (!isset($this->classes[spl_object_hash($type)])) {
             $skip = in_array($type->getSchema()->getTargetNamespace(), $this->baseSchemas, true);
@@ -298,7 +302,6 @@ class PhpConverter extends AbstractConverter
             }
 
             if (($this->isArrayType($type) || $this->isArrayNestedElement($type)) && !$force) {
-
                 $this->classes[spl_object_hash($type)]['skip'] = true;
                 $this->skipByType[spl_object_hash($class)] = true;
 
@@ -308,19 +311,15 @@ class PhpConverter extends AbstractConverter
             $this->classes[spl_object_hash($type)]['skip'] = $skip || (bool)$this->getTypeAlias($type);
         } elseif ($force) {
             if (!($type instanceof SimpleType) && !$this->getTypeAlias($type)) {
-                $this->classes[spl_object_hash($type)]['skip'] = in_array($type->getSchema()->getTargetNamespace(), $this->baseSchemas, true);
+                $this->classes[spl_object_hash($type)]['skip'] =
+                    in_array($type->getSchema()->getTargetNamespace(), $this->baseSchemas, true);
             }
         }
 
         return $this->classes[spl_object_hash($type)]['class'];
     }
 
-    /**
-     * @param string $name
-     *
-     * @return \GoetasWebservices\Xsd\XsdToPhp\Php\Structure\PHPClass
-     */
-    private function visitTypeAnonymous(Type $type, $name, PHPClass $parentClass)
+    private function visitTypeAnonymous(Type $type, string $name, PHPClass $parentClass): PHPClass
     {
         if (!isset($this->classes[spl_object_hash($type)])) {
             $this->classes[spl_object_hash($type)]['class'] = $class = new PHPClass();
@@ -340,13 +339,13 @@ class PhpConverter extends AbstractConverter
         return $this->classes[spl_object_hash($type)]['class'];
     }
 
-    private function visitComplexType(PHPClass $class, ComplexType $type)
+    private function visitComplexType(PHPClass $class, ComplexType $type): void
     {
         $schema = $type->getSchema();
         foreach ($type->getElements() as $element) {
-			if ($element instanceof Sequence) {
-				$this->visitSequence($class, $schema, $element);
-			} elseif ($element instanceof Choice) {
+            if ($element instanceof Sequence) {
+                $this->visitSequence($class, $schema, $element);
+            } elseif ($element instanceof Choice) {
                 $this->visitChoice($class, $schema, $element);
             } elseif ($element instanceof Group) {
                 $this->visitGroup($class, $schema, $element);
@@ -357,7 +356,7 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private function visitSimpleType(PHPClass $class, SimpleType $type)
+    private function visitSimpleType(PHPClass $class, SimpleType $type): void
     {
         if ($restriction = $type->getRestriction()) {
             $parent = $restriction->getBase();
@@ -387,7 +386,7 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private function handleClassExtension(PHPClass $class, Type $type)
+    private function handleClassExtension(PHPClass $class, Type $type): void
     {
         if ($alias = $this->getTypeAlias($type)) {
             $c = PHPClass::createFromFQCN($alias);
@@ -401,7 +400,7 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private function visitBaseComplexType(PHPClass $class, BaseComplexType $type)
+    private function visitBaseComplexType(PHPClass $class, BaseComplexType $type): void
     {
         $parent = $type->getParent();
         if ($parent) {
@@ -422,8 +421,12 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private function visitAttribute(PHPClass $class, Schema $schema, AttributeItem $attribute, $arrayize = true)
-    {
+    private function visitAttribute(
+        PHPClass $class,
+        Schema $schema,
+        AttributeSingle $attribute,
+        bool $arrayize = true
+    ): PHPProperty {
         $property = new PHPProperty();
         $property->setName($this->getNamingStrategy()->getPropertyName($attribute));
 
@@ -441,31 +444,38 @@ class PhpConverter extends AbstractConverter
         }
 
         $property->setDoc($attribute->getDoc());
+        if ($attribute instanceof AttributeSingle) {
+            $property->setFixed($attribute->getFixed());
+        }
 
         return $property;
     }
 
-    /**
-     * @param Element $element
-     * @param bool $arrayize
-     *
-     * @return \GoetasWebservices\Xsd\XsdToPhp\Php\Structure\PHPProperty
-     */
-    private function visitElement(PHPClass $class, Schema $schema, ElementSingle $element, $arrayize = true)
-    {
+    private function visitElement(
+        PHPClass $class,
+        Schema $schema,
+        ElementItem|ElementSingle $element,
+        bool $arrayize = true
+    ): PHPProperty {
         $property = new PHPProperty();
         $property->setName($this->getNamingStrategy()->getPropertyName($element));
         $property->setDoc($element->getDoc());
-        if ($element->isNil() || $element->getMin() === 0) {
+
+        if (
+            ($element instanceof ElementSingle && $element->isNil()) ||
+            ($element instanceof InterfaceSetMinMax && $element->getMin() === 0)
+        ) {
             $property->setNullable(true);
+        }
+
+        if (!$element instanceof ElementSingle) {
+            return $property;
         }
 
         $t = $element->getType();
 
         if ($arrayize) {
-
             if ($itemOfArray = $this->isArrayType($t)) {
-
                 if (!$itemOfArray->getName()) {
                     if ($element instanceof ElementRef) {
                         $refClass = $this->visitElementDef($element->getReferencedElement());
@@ -485,7 +495,6 @@ class PhpConverter extends AbstractConverter
 
                 return $property;
             } elseif ($itemOfArray = $this->isArrayNestedElement($t)) {
-
                 if (!$t->getName()) {
                     if ($element instanceof ElementRef) {
                         $refClass = $this->visitElementDef($element->getReferencedElement());
@@ -514,11 +523,14 @@ class PhpConverter extends AbstractConverter
         }
 
         $property->setType($this->findPHPElementClassName($class, $element));
+        if ($element instanceof InterfaceSetFixed) {
+            $property->setFixed($element->getFixed());
+        }
 
         return $property;
     }
 
-    private function findPHPClass(PHPClass $class, Item $node, $force = false)
+    private function findPHPClass(PHPClass $class, Item $node, bool $force = false): PHPClass
     {
         if ($node instanceof ElementRef) {
             return $this->visitElementDef($node->getReferencedElement());
@@ -533,7 +545,7 @@ class PhpConverter extends AbstractConverter
         }
     }
 
-    private function typeHasValue(Type $type, PHPClass $parentClass, $name)
+    private function typeHasValue(Type $type, PHPClass $parentClass, string $name): PHPClass|false
     {
         $newType = null;
         do {
@@ -567,13 +579,14 @@ class PhpConverter extends AbstractConverter
         return false;
     }
 
-    private function findPHPElementClassName(PHPClass $class, ElementItem $element)
+    private function findPHPElementClassName(PHPClass $class, ElementItem $element): PHPClass
     {
         if ($element instanceof ElementRef) {
-            $elRefClass = $this->visitElementDef($element->getReferencedElement());
-            $refType = $this->findPHPClass($elRefClass, $element->getReferencedElement());
+            $elRef = $element->getReferencedElement();
+            $elRefClass = $this->visitElementDef($elRef);
+            $refType = $this->findPHPClass($elRefClass, $elRef);
 
-            if ($this->typeHasValue($element->getReferencedElement()->getType(), $elRefClass, $element->getReferencedElement())) {
+            if ($this->typeHasValue($elRef->getType(), $elRefClass, $elRef->getName())) {
                 return $refType;
             }
         }
