@@ -1,54 +1,17 @@
 <?php
 
-namespace GoetasWebservices\Xsd\XsdToPhp\Tests\JmsSerializer\OTA;
+namespace GoetasWebservices\Xsd\XsdToPhp\Tests\PHP;
 
-use GoetasWebservices\XML\XSDReader\SchemaReader;
-use GoetasWebservices\Xsd\XsdToPhp\Naming\ShortNamingStrategy;
-use GoetasWebservices\Xsd\XsdToPhp\Php\ClassGenerator;
-use GoetasWebservices\Xsd\XsdToPhp\Php\PhpConverter;
 use PHPUnit\Framework\TestCase;
 
 class PHPConversionTest extends TestCase
 {
-    /**
-     * @param mixed $xml
-     *
-     * @return \Laminas\Code\Generator\ClassGenerator[]
-     */
-    protected function getClasses($xml)
-    {
-        $phpcreator = new PhpConverter(new ShortNamingStrategy());
-        $phpcreator->addNamespace('http://www.example.com', 'Example');
+    use GetPhpYamlTrait;
 
-        $generator = new ClassGenerator();
-        $reader = new SchemaReader();
-
-        if (!is_array($xml)) {
-            $xml = [
-                'schema.xsd' => $xml,
-            ];
-        }
-        $schemas = [];
-        foreach ($xml as $name => $str) {
-            $schemas[] = $reader->readString($str, $name);
-        }
-        $items = $phpcreator->convert($schemas);
-
-        $classes = [];
-        foreach ($items as $k => $item) {
-            if ($codegen = $generator->generate($item)) {
-                $classes[$k] = $codegen;
-            }
-        }
-
-        return $classes;
-    }
-
-    public function testSimpleContent()
+    public function testSimpleContent(): void
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com"
-            xmlns:xs="http://www.w3.org/2001/XMLSchema">
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="single">
                     <xs:simpleContent>
     					<xs:extension base="xs:string">
@@ -58,7 +21,7 @@ class PHPConversionTest extends TestCase
                 </xs:complexType>
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
         $this->assertCount(1, $items);
 
         $codegen = $items['Example\SingleType'];
@@ -71,11 +34,10 @@ class PHPConversionTest extends TestCase
         $this->assertTrue($codegen->hasMethod('setCode'));
     }
 
-    public function testSimpleNoAttributesContent()
+    public function testSimpleNoAttributesContent(): void
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com"
-            xmlns:xs="http://www.w3.org/2001/XMLSchema">
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="single">
                     <xs:simpleContent>
     					<xs:extension base="xs:string"/>
@@ -86,7 +48,7 @@ class PHPConversionTest extends TestCase
                 </xs:simpleType>
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
         $this->assertCount(1, $items);
 
         $codegen = $items['Example\SingleType'];
@@ -96,11 +58,10 @@ class PHPConversionTest extends TestCase
         $this->assertTrue($codegen->hasMethod('__toString'));
     }
 
-    public function testNoMulteplicity()
+    public function testNoMulteplicity(): void
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com"
-            xmlns:xs="http://www.w3.org/2001/XMLSchema">
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="single">
                     <xs:all>
                         <xs:element name="id" type="xs:long" minOccurs="0"/>
@@ -108,7 +69,7 @@ class PHPConversionTest extends TestCase
                 </xs:complexType>
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
         $this->assertCount(1, $items);
 
         $codegen = $items['Example\SingleType'];
@@ -119,11 +80,10 @@ class PHPConversionTest extends TestCase
         $this->assertTrue($codegen->hasMethod('setId'));
     }
 
-    public function testMulteplicity()
+    public function testMulteplicity(): void
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com"
-            xmlns:xs="http://www.w3.org/2001/XMLSchema">
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="single">
                     <xs:all>
                         <xs:element name="id" type="ary" minOccurs="0"/>
@@ -136,7 +96,7 @@ class PHPConversionTest extends TestCase
                 </xs:complexType>
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
 
         $this->assertCount(2, $items);
 
@@ -151,10 +111,10 @@ class PHPConversionTest extends TestCase
         $this->assertNull($codegen->getMethod('issetId')->getParameters()['index']->getType());
     }
 
-    public function testNestedMulteplicity()
+    public function testNestedMulteplicity(): void
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com" xmlns:xs="http://www.w3.org/2001/XMLSchema">
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="single">
                     <xs:all>
                         <xs:element name="id" type="ary" minOccurs="0"/>
@@ -172,7 +132,7 @@ class PHPConversionTest extends TestCase
                 </xs:complexType>
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
 
         $this->assertCount(3, $items);
 
@@ -197,25 +157,21 @@ class PHPConversionTest extends TestCase
     public function testMultipleArrayTypes()
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com"
-            xmlns:xs="http://www.w3.org/2001/XMLSchema">
-
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="ArrayOfStrings">
                     <xs:all>
                         <xs:element name="string" type="xs:string" maxOccurs="unbounded"/>
                     </xs:all>
                 </xs:complexType>
-
                 <xs:complexType name="Single">
                     <xs:all>
                         <xs:element name="a" type="ArrayOfStrings"/>
                         <xs:element name="b" type="ArrayOfStrings"/>
                     </xs:all>
                 </xs:complexType>
-
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
 
         $this->assertCount(2, $items);
 
@@ -231,8 +187,7 @@ class PHPConversionTest extends TestCase
     public function testSimpleMulteplicity()
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com"
-            xmlns:xs="http://www.w3.org/2001/XMLSchema">
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="single">
                     <xs:all>
                         <xs:element name="id" type="ary" minOccurs="0"/>
@@ -243,7 +198,7 @@ class PHPConversionTest extends TestCase
                 </xs:simpleType>
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
 
         $this->assertCount(1, $items);
 
@@ -258,8 +213,7 @@ class PHPConversionTest extends TestCase
     public function testNillableElement()
     {
         $xml = '
-            <xs:schema targetNamespace="http://www.example.com"
-            xmlns:xs="http://www.w3.org/2001/XMLSchema">
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
                 <xs:complexType name="single">
                     <xs:all>
                         <xs:element name="date1" type="xs:date" nillable="true"/>
@@ -270,16 +224,76 @@ class PHPConversionTest extends TestCase
                 </xs:complexType>
             </xs:schema>';
 
-        $items = $this->getClasses($xml);
+        $items = $this->getPhpClasses($xml);
         $codegen = $items['Example\SingleType'];
 
         $this->assertTrue($codegen->hasMethod('setDate1'));
-        $this->assertNull($codegen->getMethod('setDate1')->getParameters()['date1']->getDefaultValue()->getValue());
+        $this->assertNull($codegen->getMethod('setDate1')->getParameters()['date1']->getDefaultValue());
         $this->assertTrue($codegen->hasMethod('setDate2'));
         $this->assertNull($codegen->getMethod('setDate2')->getParameters()['date2']->getDefaultValue());
         $this->assertTrue($codegen->hasMethod('setStr1'));
         $this->assertNull($codegen->getMethod('setStr1')->getParameters()['str1']->getDefaultValue());
         $this->assertTrue($codegen->hasMethod('setStr2'));
         $this->assertNull($codegen->getMethod('setStr2')->getParameters()['str2']->getDefaultValue());
+    }
+
+    public function testNoCodeDuplicationInExtendingClass(): void
+    {
+        $xml = '
+            <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
+                <xs:complexType name="Code">
+                    <xs:sequence>
+                        <xs:element name="code" type="xs:token" form="unqualified"/>
+                        <xs:element name="name" type="xs:normalizedString" form="unqualified" fixed="foo"/>
+                    </xs:sequence>
+                    <xs:attribute name="listURI" type="xs:anyURI" use="optional"/>
+                    <xs:attribute name="listVersionID" type="xs:normalizedString" use="optional"/>
+                </xs:complexType>
+                <xs:simpleType name="Codeliste.Zeitserientyp">
+                    <xs:restriction base="xs:token">
+                      <xs:enumeration value="01"/>
+                      <xs:enumeration value="02"/>
+                    </xs:restriction>
+                </xs:simpleType>
+                <xs:complexType name="Code.Zeitserientyp">
+                    <xs:complexContent>
+                      <xs:restriction base="Code">
+                        <xs:sequence>
+                          <xs:element name="code" type="Codeliste.Zeitserientyp" form="unqualified"/>
+                        </xs:sequence>
+                        <xs:attribute name="listURI" type="xs:anyURI" use="optional" fixed="urn:xoev-de:fim:codeliste:xzufi.zeitserientyp"/>
+                        <xs:attribute name="listVersionID" type="xs:normalizedString" use="optional" fixed="1.1"/>
+                      </xs:restriction>
+                    </xs:complexContent>
+                  </xs:complexType>
+            </xs:schema>';
+
+        $items = $this->getPhpClasses($xml);
+        $codeType = $items['Example\CodeType'];
+        $codeZeitserientypType = $items['Example\CodeZeitserientypType'];
+
+        $this->assertTrue($codeType->hasMethod('setCode'));
+        $this->assertTrue($codeType->hasMethod('getCode'));
+        $this->assertTrue($codeType->hasMethod('setName'));
+        $this->assertTrue($codeType->hasMethod('getName'));
+        $this->assertTrue($codeType->hasMethod('setListURI'));
+        $this->assertTrue($codeType->hasMethod('getListURI'));
+        $this->assertNull($codeType->getProperty('listURI')->getDefaultValue());
+        $this->assertTrue($codeType->hasMethod('setListVersionID'));
+        $this->assertTrue($codeType->hasMethod('getListVersionID'));
+        $this->assertNull($codeType->getProperty('listVersionID')->getDefaultValue());
+        $this->assertTrue($codeZeitserientypType->getExtendedClass() === 'Example\CodeType');
+        $this->assertFalse($codeZeitserientypType->hasMethod('setCode'));
+        $this->assertFalse($codeZeitserientypType->hasMethod('getCode'));
+        $this->assertFalse($codeZeitserientypType->hasMethod('setName'));
+        $this->assertFalse($codeZeitserientypType->hasMethod('getName'));
+        $this->assertFalse($codeZeitserientypType->hasMethod('setListURI'));
+        $this->assertFalse($codeZeitserientypType->hasMethod('getListURI'));
+        $value = $codeZeitserientypType->getProperty('listURI')->getDefaultValue()->getValue();
+        $this->assertTrue($value === "urn:xoev-de:fim:codeliste:xzufi.zeitserientyp");
+        $this->assertFalse($codeZeitserientypType->hasMethod('setListVersionID'));
+        $this->assertFalse($codeZeitserientypType->hasMethod('getListVersionID'));
+        $value = $codeZeitserientypType->getProperty('listVersionID')->getDefaultValue()->getValue();
+        $this->assertTrue($value === "1.1");
     }
 }
